Word Count complete
We now have the interpreter creating a graph of the html. The currently
only has a word count but we know have the structure to make more
interpretation of the html.
We are ready to add weight to certain html elements (we give a weight
of zero to irrelevant elements such as script)
Weights are supplied in the admin of the django application.
This will currently make a db query for every request - should have way
to cache the request.